refactor unitConversion.js
